feat: question content conditional on whether the question is answered
Added logic in the question component to check whether the authed user already
has answered the question using local state. If the question is answered the
question stats are shown and the answer button is disabled. If the question is
not answered yet the question stats are not shown and the user can answer the
question. Next step is including the dispatch to answer the question