feat: handle answering of a question
Added the api call to answer the question into the question action creator and
a handleAnswerQuestion in the component itself. Also had to update the
reducer function a little bit to correctly change the store. As outlined in the
project description the update to the database does not happen on the same time
as a result although the question is answered it is not show immediately on the
page. Next step is updating the homepage when a question is answered