fix: add user actions to deal with new and answered questions
The user action creator and reducer did not account for questions added or
answered by the user as such the user part of the store was never updated
and the homepage with the answered and unanswered question did to show the
correct questions. Added an save question and save answer action to the user
action creator and updated the user reducer to reflect the new store.
Dispatching is handled in the question action creators. Additionally added
a simple 404 page