fix: fixed question component and user reducer
The question component was not rendering the 404 component in case a user
type a non-existing question id in the address bar additionally the poll data
was not displayed correctly when user answered the poll. Changes consist of a
if statement in the render checking if the question id exist if not display
the 404 component and instead of using componentdidmount and using a local state use the store data directly. The user reducer was not strutured correctly and
as a result when a user created a new question a type error occured. Corrected
the structure of the store update